requirements.txt
annotated-types==0.7.0
anyio==4.12.1
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
cryptography==46.0.5
cssselect2==0.9.0
distro==1.9.0
Django==6.0.3
django-markdownify==0.9.6
freetype-py==2.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
lxml==6.0.2
Markdown==3.10.2
oscrypto==1.3.0
openpyxl==3.1.5
pillow==12.1.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyHanko==0.34.1
pyhanko-certvalidator==0.30.1
pyparsing==3.3.2
pypdf==6.9.1
python-bidi==0.6.7
python-dotenv==1.2.2
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
svglib==1.6.0
tenacity==9.1.4
tinycss2==1.4.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
uritools==6.0.1
urllib3==2.6.3
webencodings==0.5.1
websockets==16.0
xhtml2pdf==0.2.17
